getOmniProtocolUrlMap
