+++ features/omni-kit/helpers/index.ts
@@ -25,3 +25,4 @@ export * from './resolvePaybackAmount'
 export * from './mapOmniLendingCumulatives'
 export * from './mapOmniEarnCumulatives'
 export * from './mapBorrowCumulativesToOmniCumulatives'
+export * from './getOmniProtocolUrlMap'
